+++ lab5/Bai_tap_2/main.cpp
@@ -9,13 +9,13 @@ int main(){
 
     // tong so luong cho tung loai giao dich
     C1.So_luong();
-    cout<<"Tien trung binh cua giao dich can ho chung cu la: "<<C1.thanhtienTB()<<" dong"<<endl;
+    cout<<"Tien trung binh cua giao dich can ho chung cu la: "<<(long long)C1.thanhtienTB()<<" dong"<<endl;
     nhapho* caonhat = C1.max_tien();
     if(!caonhat){
     cout << "Khong co giao dich nha pho" << endl;
     } 
     else {
-        cout << "Thanh tien lon nhat: " << caonhat->Thanhtien() << endl;
+        cout << "Thanh tien lon nhat: " << (long long)caonhat->Thanhtien() << endl;
         caonhat->xuat();
     }
     vector<giaodich*> gd12 = C1.thang12();
